requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pendulum==1.4.4
psycopg2==2.7.4
python-dateutil==2.7.2
python-telegram-bot==10.0.1
pytz==2018.3
pytzdata==2018.3
requests==2.18.4
six==1.11.0
tzlocal==1.5.1
urllib3==1.22
Werkzeug==0.14.1
